pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "elastihone"
version = "0.1.0"
description = "AI-powered detection rule fine-tuning for Elastic Security"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "ElastiHone Contributors" }]
keywords = ["siem", "detection", "sigma", "elastic", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "pydantic-ai>=0.1",
    "elasticsearch[async]>=8.17,<9",
    "pySigma>=0.11",
    "pySigma-backend-elasticsearch>=1.2",
    "mcp>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "typer[all]>=0.15",
    "rich>=13.9",
    "pyyaml>=6.0",
    "httpx>=0.28",
    "openai>=1.60",
    "aiosqlite>=0.20",
    "slowapi>=0.1.9",
    "cryptography>=44.0",
    "markupsafe>=2.1",
    "markdown-it-py>=3.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
    "pytest-cov>=6.0",
]

[project.scripts]
sda = "sda.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/sda"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
